Changes made when putting the robot together and splines

+++ Robot/Wheel.cpp
@@ -41,7 +41,8 @@ void Wheel::drive(short vel){
       digitalWrite(pin_in1, LOW);
       digitalWrite(pin_in2, HIGH);
     }
-    analogWrite(pin_en,vel);
+//    Serial.println((byte)vel);
+    analogWrite(pin_en, (byte)vel);
 }
 
 void Wheel::brake(){
